\name{as_alpha_pairwise}
\alias{as_alpha_pairwise}
\title{Extract alpha diversity pairwise tests}
\usage{
as_alpha_pairwise(x)
}
\arguments{
\item{x}{A \code{microeda_alpha_compare} object.}
}
\value{
A data frame of pairwise Wilcoxon test results.
}
\description{
Extract the pairwise alpha-diversity comparison table from
\code{microeda_alpha_compare()}.
}
