\name{as_alpha_tests}
\alias{as_alpha_tests}
\title{Extract alpha diversity omnibus tests}
\usage{
as_alpha_tests(x)
}
\arguments{
\item{x}{A \code{microeda_alpha_compare} object.}
}
\value{
A data frame of Kruskal-Wallis test results.
}
\description{
Extract the omnibus alpha-diversity comparison table from
\code{microeda_alpha_compare()}.
}
