% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-ordination.R
\name{as_beta_coordinates}
\alias{as_beta_coordinates}
\title{Extract beta ordination coordinates}
\usage{
as_beta_coordinates(x)
}
\arguments{
\item{x}{A \code{microeda_beta_ordination} object.}
}
\value{
A data frame with one row per sample and one column per ordination
axis.
}
\description{
Extract beta ordination coordinates
}
