% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{as_beta_dist}
\alias{as_beta_dist}
\title{Extract beta diversity distances}
\usage{
as_beta_dist(x)
}
\arguments{
\item{x}{A \code{microeda_beta} object.}
}
\value{
A \code{\link[stats:dist]{stats::dist()}} object.
}
\description{
Extract beta diversity distances
}
