% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{as_beta_matrix}
\alias{as_beta_matrix}
\title{Extract beta diversity distances as a matrix}
\usage{
as_beta_matrix(x)
}
\arguments{
\item{x}{A \code{microeda_beta} object.}
}
\value{
A square numeric matrix of beta diversity distances.
}
\description{
Extract beta diversity distances as a matrix
}
