% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{as_beta_samples}
\alias{as_beta_samples}
\title{Extract beta diversity sample labels}
\usage{
as_beta_samples(x)
}
\arguments{
\item{x}{A \code{microeda_beta} object.}
}
\value{
A data frame with \code{sample_id} and, when present, \code{group}.
}
\description{
Extract beta diversity sample labels
}
