% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-test.R
\name{as_beta_test_summary}
\alias{as_beta_test_summary}
\title{Summarize beta group test results}
\usage{
as_beta_test_summary(x)
}
\arguments{
\item{x}{A \code{microeda_beta_test} object.}
}
\value{
A one-row data frame with compact PERMANOVA and dispersion columns.
}
\description{
Summarize beta group test results
}
