% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microEDA-class.R
\docType{methods}
\name{checkProfileRank}
\alias{checkProfileRank}
\alias{checkProfileRank,microEDA-method}
\title{Check taxonomic rank of the profile}
\usage{
checkProfileRank(object)

\S4method{checkProfileRank}{microEDA}(object)
}
\arguments{
\item{object}{An object of class \code{microEDA}.}
}
\value{
The taxonomic rank stored in the \code{info} slot.
}
\description{
Check taxonomic rank of the profile
}
