% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microEDA-class.R
\docType{methods}
\name{getSampleData}
\alias{getSampleData}
\alias{getSampleData,microEDA-method}
\title{Extract sample data}
\usage{
getSampleData(object)

\S4method{getSampleData}{microEDA}(object)
}
\arguments{
\item{object}{An object of class \code{microEDA}.}
}
\value{
The sam_data component of the microEDA/phyloseq object.
}
\description{
Extract sample data
}
