% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microEDA-class.R
\docType{methods}
\name{getTaxTable}
\alias{getTaxTable}
\alias{getTaxTable,microEDA-method}
\title{Extract taxonomic table}
\usage{
getTaxTable(object)

\S4method{getTaxTable}{microEDA}(object)
}
\arguments{
\item{object}{An object of class \code{microEDA}.}
}
\value{
The tax_table component of the microEDA/phyloseq object.
}
\description{
Extract taxonomic table
}
