% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlanProfile-class.R
\name{load_metaphlan}
\alias{load_metaphlan}
\title{Load a MetaPhlAn taxonomic profile from a file}
\usage{
load_metaphlan(file_path)
}
\arguments{
\item{file_path}{A \code{character} string containing the path to the file to read.}
}
\value{
An object of class \code{metaphlanProfile}.
}
\description{
Reads a tab-separated MetaPhlAn output file and constructs a \linkS4class{metaphlanProfile} object.
The input file should contain taxonomic abundance data with columns for clade name and sample abundances.
}
\examples{
file_path <- system.file("extdata", "merged_abundance_table.txt", package = "microEDA")

mtp <- load_metaphlan(file_path)
}
