% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{merged_metaphlan_profiles}
\alias{merged_metaphlan_profiles}
\title{Merged MetaPhlAn4 profiles data}
\format{
\subsection{\code{merged_abundance_profiles}}{

An object of class \linkS4class{metaphlanProfile} with slots \code{mpa_version} and \code{mpa_taxProfile}:
\describe{
\item{mpa_version}{A \code{Character} vector indicating the MetaPhlAn database version used.}
\item{mpa_taxProfile}{A data frame with 59 rows and 7 columns:

\code{clade_name}: The taxonomic lineage of the taxon reported on this row.
Taxon names are prefixed with one-letter codes to help indicate their rank.

\code{SRS0144...}: The sample names for which sample-specific abundance
profiles are found in a column.}
}
}
}
\source{
\url{https://github.com/biobakery/MetaPhlAn/releases/tag/4.1.0}
}
\usage{
merged_metaphlan_profiles
}
\description{
Example data of multiple MetaPhlAn4 profiles that were merged into one file
in \code{metaphlanProfile} format.
}
\keyword{datasets}
