% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlanProfile-class.R
\docType{class}
\name{metaphlanProfile-class}
\alias{metaphlanProfile-class}
\title{A class to represent a MetaPhlAn profile.}
\description{
This class stores database version information and taxonomic profile data from MetaPhlAn output.
}
\section{Slots}{

\describe{
\item{\code{mpa_version}}{\code{Character} string indicating the MetaPhlAn database version used.}

\item{\code{mpa_taxProfile}}{A \code{data.frame} containing the taxonomic profile (relative abundances of clades).}
}}

