% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microEDA-class.R
\docType{class}
\name{microEDA-class}
\alias{microEDA-class}
\alias{microEDA}
\alias{microEDA,metaphlanProfile-method}
\alias{microEDA,phyloseq-method}
\alias{show,microEDA-method}
\title{A S4 class extending phyloseq for microbiome data analysis}
\usage{
microEDA(tax_profile, metadata = NULL, sample_column = NULL)

\S4method{microEDA}{metaphlanProfile}(tax_profile, metadata = NULL, sample_column = NULL)

\S4method{microEDA}{phyloseq}(tax_profile, metadata = NULL, sample_column = NULL)

\S4method{show}{microEDA}(object)
}
\arguments{
\item{tax_profile}{An object of class \code{metaphlanProfile} or \code{phyloseq}.}

\item{metadata}{Optional \code{data.frame} with sample metadata.}

\item{sample_column}{\code{Character} string specifying the column name in
\code{metadata} that contains sample identifiers (e.g., "Samples").}

\item{object}{A \code{microEDA} object.}
}
\value{
An object of class \code{microEDA}.
}
\description{
This class inherits from \code{\link[phyloseq]{phyloseq}} and adds a slot for
additional information, such as lowest taxonomic rank, MetaPhlAn database
version, applied transformations and/or filters.

Displays a summary of a \code{microEDA} object, including the standard
\code{phyloseq} output and custom information from the \code{info} slot.
}
\details{
When passed a \linkS4class{metaphlanProfile}, the profile will be returned at
the lowest available taxonomic rank, which usually is at strain rank. To
agglomerate the profile at a higher rank, please see \code{\link[=agglomerate_taxa]{agglomerate_taxa()}}.
}
\section{Slots}{

\describe{
\item{\code{info}}{A \code{list} containing metadata.}

\item{\code{otu_table}}{A \code{matrix} of class \code{otu_table} with taxa as rows and samples as columns.}

\item{\code{tax_table}}{A named \verb{character matrix} of class \code{taxonomyTable}, with taxon IDs as row names and taxonomic ranks as columns.}

\item{\code{sam_data}}{An object of class \code{sample_data} containing sample metadata.}

\item{\code{phy_tree}}{An object of class \code{phylo} representing a phylogenetic tree.}

\item{\code{refseq}}{An object of a class inheriting from \code{XStringSet-class} (e.g., \code{DNAStringSet}) containing biological sequences. See \code{\link[Biostrings]{XStringSet}}.}
}}

\examples{
data("GlobalPatterns", package = "phyloseq")
me <- microEDA(GlobalPatterns)
}
