% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microEDA-class.R
\name{otu_table<-,microEDA,otu_table-method}
\alias{otu_table<-,microEDA,otu_table-method}
\title{Replace OTU table in a microEDA object}
\usage{
\S4method{otu_table}{microEDA,otu_table}(x) <- value
}
\arguments{
\item{x}{A \code{microEDA} object.}

\item{value}{An \code{otu_table} object (or coercible to one).}
}
\value{
The updated \code{microEDA} object.
}
\description{
Replace OTU table in a microEDA object
}
