% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microEDA-class.R
\name{phy_tree<-,microEDA,phylo-method}
\alias{phy_tree<-,microEDA,phylo-method}
\title{Replace phylogenetic tree in a microEDA object}
\usage{
\S4method{phy_tree}{microEDA,phylo}(x) <- value
}
\arguments{
\item{x}{A \code{microEDA} object.}

\item{value}{A \code{phylo} object.}
}
\value{
The updated \code{microEDA} object.
}
\description{
Replace phylogenetic tree in a microEDA object
}
