% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microEDA-class.R
\name{sample_data<-}
\alias{sample_data<-}
\alias{sample_data<-,microEDA,sample_data-method}
\title{Replace sample data in a microEDA object}
\usage{
sample_data(x) <- value
}
\arguments{
\item{x}{A \code{microEDA} object.}

\item{value}{A \code{sample_data} object (or coercible to one).}
}
\value{
The updated \code{microEDA} object.
}
\description{
Replace sample data in a microEDA object
}
