% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{single_metaphlan_profile}
\alias{single_metaphlan_profile}
\title{SRS014476-Supragingival_plaque data}
\format{
\subsection{\code{single_abundance_profile}}{

An object of class \linkS4class{metaphlanProfile} with slots \code{mpa_version} and \code{mpa_taxProfile}:
\describe{
\item{mpa_version}{A \code{Character} vector indicating the MetaPhlAn database version used.}
\item{mpa_taxProfile}{A data frame with 20 rows and 2 columns:

\code{clade_name}: The taxonomic lineage of the taxon reported on this row.
Taxon names are prefixed with one-letter codes to help indicate their rank.

\code{single_sample_profile}: Each taxon's relative abundance in \%.}
}
}
}
\source{
\url{https://github.com/biobakery/MetaPhlAn/releases/tag/4.1.0}
}
\usage{
single_metaphlan_profile
}
\description{
Example data of a single MetaPhlAn4 output file in \code{metaphlanProfile}
format, containing the computed taxon abundances.
}
\keyword{datasets}
