% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microEDA-class.R
\name{summary,microEDA-method}
\alias{summary,microEDA-method}
\title{Summarize a microEDA object}
\usage{
\S4method{summary}{microEDA}(object, ...)
}
\arguments{
\item{object}{A \code{microEDA} object.}

\item{...}{Optional arguments (currently supports \code{label_width}).}
}
\value{
An invisible list containing summary statistics.
}
\description{
Provides a concise, more detailed overview over a \code{microEDA} object than
the \link[=microEDA-class]{show()} method.
}
