% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microEDA-class.R
\name{tax_table<-,microEDA,taxonomyTable-method}
\alias{tax_table<-,microEDA,taxonomyTable-method}
\title{Replace taxonomic table in a microEDA object}
\usage{
\S4method{tax_table}{microEDA,taxonomyTable}(x) <- value
}
\arguments{
\item{x}{A \code{microEDA} object.}

\item{value}{A \code{taxonomyTable} object (or coercible to one).}
}
\value{
The updated \code{microEDA} object.
}
\description{
Replace taxonomic table in a microEDA object
}
