% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_phyloseq-methods.R
\name{to_phyloseq}
\alias{to_phyloseq}
\alias{to_phyloseq,microEDA-method}
\alias{to_phyloseq,metaphlanProfile-method}
\title{Convert an object to a phyloseq object}
\usage{
to_phyloseq(x, ...)

\S4method{to_phyloseq}{microEDA}(x)

\S4method{to_phyloseq}{metaphlanProfile}(x, metadata = NULL, sample_column = NULL)
}
\arguments{
\item{x}{An object of class \code{microEDA}.}

\item{...}{Additional arguments passed to methods (e.g., \code{metadata}, \code{sample_column}).}

\item{metadata}{Optional \code{data.frame} containing sample metadata.}

\item{sample_column}{\code{Character} string specifying the column in \code{metadata} with sample IDs.}
}
\value{
An object of class \code{phyloseq}.
}
\description{
S4 generic to convert objects into a \code{phyloseq} object.
Currently, methods for \code{metaphlanProfile} and \code{microEDA} objects are implemented.
}
\examples{
# Convert a microEDA object to phyloseq
data(enterotype, package = "phyloseq")
me <- microEDA(enterotype)
is(me, "microEDA")
ps <- to_phyloseq(me)
is(ps, "phyloseq")

# Convert a metaphlanProfile object to phyloseq
mpa <- microEDA::merged_metaphlan_profiles
is(mpa, "metaphlanProfile")
ps <- to_phyloseq(mpa)
is(ps, "phyloseq")
}
\seealso{
\code{\link[phyloseq]{phyloseq}}, \code{\link{microEDA}}, \code{\link{metaphlanProfile-class}}
}
