% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{as_alpha_pairwise}
\alias{as_alpha_pairwise}
\title{Extract alpha diversity pairwise tests}
\usage{
as_alpha_pairwise(x)
}
\arguments{
\item{x}{A \code{microeda_alpha_compare} object.}
}
\value{
A data frame of pairwise Wilcoxon test results, including the
Wilcoxon test statistic.
}
\description{
Extract the pairwise alpha-diversity comparison table from
\code{\link[=microeda_alpha_compare]{microeda_alpha_compare()}}.
}
