\name{as_alpha_summary}
\alias{as_alpha_summary}
\title{Extract alpha diversity group summaries}
\usage{
as_alpha_summary(x)
}
\arguments{
\item{x}{A \code{microeda_alpha} object.}
}
\value{
A data frame with summary statistics by group and index.
}
\description{
Extract group-level summaries from \code{microeda_alpha()}.
}
