% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{as_alpha_table}
\alias{as_alpha_table}
\title{Extract per-sample alpha diversity indices}
\usage{
as_alpha_table(x)
}
\arguments{
\item{x}{A \code{microeda_alpha} object.}
}
\value{
A data frame with one row per sample.
}
\description{
Extract the per-sample alpha diversity table from \code{\link[=microeda_alpha]{microeda_alpha()}}.
}
