\name{as_recommendations}
\alias{as_recommendations}
\title{Extract recommendations from a microeda report}
\usage{
as_recommendations(x)
}
\arguments{
\item{x}{A \code{microeda_report} object.}
}
\value{
A data frame of recommendations.
}
\description{
Extract the recommendations table from the result of \code{microeda_check()}.
}
