% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microeda.R
\name{as_recommendations}
\alias{as_recommendations}
\title{Extract recommendations from a microeda report}
\usage{
as_recommendations(x)
}
\arguments{
\item{x}{A \code{microeda_report} object.}
}
\value{
A data frame of recommendations.
}
\description{
Extract the recommendations table from the result of \code{\link[=microeda_check]{microeda_check()}}.
}
