% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microEDA-class.R
\docType{methods}
\name{info}
\alias{info}
\alias{info,microEDA-method}
\alias{info<-}
\alias{info<-,microEDA-method}
\alias{taxrank}
\alias{taxrank,microEDA-method}
\alias{taxrank<-}
\alias{taxrank<-,microEDA-method}
\alias{transforms}
\alias{transforms,microEDA-method}
\alias{transforms<-}
\alias{transforms<-,microEDA-method}
\alias{filters}
\alias{filters,microEDA-method}
\alias{filters<-}
\alias{filters<-,microEDA-method}
\alias{mpa_version,microEDA-method}
\alias{mpa_version<-,microEDA-method}
\title{Get or set the info slot from a microEDA object}
\usage{
info(object)

info(object) <- value

taxrank(object)

taxrank(object) <- value

transforms(object)

transforms(object) <- value

filters(object)

filters(object) <- value

\S4method{mpa_version}{microEDA}(object)

\S4method{mpa_version}{microEDA}(object) <- value
}
\arguments{
\item{object}{A \linkS4class{microEDA} object.}

\item{value}{For \verb{info<-}: A \code{list} containing metadata. Valid metadata keys are: 'taxrank', 'transforms', 'filters', 'mpa_version'.

For \verb{taxrank<-}: A \code{character} string specifying the taxonomic rank (e.g., "Species", "Genus").

For \verb{transforms<-}: A \code{character} vector of applied transformation(s) (e.g., c("TSS", "log2")).

For \verb{filters<-}: A named vector specifying used filter parameters (e.g., c(min_abundance = 0.01, min_prevalence = 0.5)).

For \verb{mpa_version<-}: A \code{character} string specifying the MetaPhlAn version (e.g., "#mpa_vJun23_CHOCOPhlAnSGB_202307").}
}
\value{
For \code{info()}: A \code{list} containing metadata. For \verb{info<-}: updated \code{microEDA} object.

For \code{taxrank()}: \code{character} string. For \verb{taxrank<-}: updated \code{microEDA} object.

For \code{transforms()}: \code{character} vector. For \verb{transforms<-}: updated \code{microEDA} object.

For \code{filters()}: list. For \verb{filters<-}: updated \code{microEDA} object.

For \code{mpa_version()}: \code{character} string. For \verb{mpa_version<-}: updated \code{microEDA} object.
}
\description{
Accessor methods for either retrieving or setting the info slot in a \code{microEDA} object.
Additionally, each expected element of the info slot can be retrieved or set
individually.
}
\details{
\code{taxrank}: represents the lowest available taxonomic rank.

\code{transforms}: represents transformation(s) that were applied to the abundance data (otu_table).

\code{filters}: represents filters that were applied to the abundance data (otu_table).

\code{mpa_version}: represents the MetaPhlAn database version that was used to create the profile.
}
\seealso{
\code{\link{mpa_version,metaphlanProfile-method}}
}
