% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microEDA-class.R
\docType{methods}
\name{info}
\alias{info}
\alias{info,microEDA-method}
\alias{info<-}
\alias{info<-,microEDA-method}
\alias{infoKeys}
\alias{infoFields}
\alias{infoFields,microEDA-method}
\alias{taxrank}
\alias{taxrank,microEDA-method}
\alias{taxrank<-}
\alias{taxrank<-,microEDA-method}
\alias{transforms}
\alias{transforms,microEDA-method}
\alias{transforms<-}
\alias{transforms<-,microEDA-method}
\alias{mpa_version,microEDA-method}
\alias{mpa_version<-,microEDA-method}
\alias{filtered_taxa}
\alias{filtered_taxa,microEDA-method}
\alias{filtered_taxa<-}
\alias{filtered_taxa<-,microEDA-method}
\alias{filters}
\alias{filters,microEDA-method}
\alias{filters<-}
\alias{filters<-,microEDA-method}
\title{Get or set the info slot from a microEDA object}
\usage{
info(object)

info(object) <- value

infoKeys()

infoFields(object)

taxrank(object)

taxrank(object) <- value

transforms(object)

transforms(object) <- value

\S4method{mpa_version}{microEDA}(object)

\S4method{mpa_version}{microEDA}(object) <- value

filtered_taxa(object)

filtered_taxa(object) <- value

filters(object)

filters(object) <- value
}
\arguments{
\item{object}{A \linkS4class{microEDA} object.}

\item{value}{Only used in replacement methods (setters).
\describe{
\item{For \verb{info<-}: }{A \code{list} containing metadata. Valid metadata keys are: 'taxrank', 'transforms', 'filters', 'mpa_version'.}
\item{For \verb{taxrank<-}: }{A \code{character} string specifying the taxonomic rank (e.g., "Species", "Genus").}
\item{For \verb{transforms<-}: }{A \code{character} vector of applied transformation(s) (e.g., c("TSS", "log2")).}
\item{For \verb{mpa_version<-}: }{A \code{character} string specifying the MetaPhlAn version (e.g., "#mpa_vJun23_CHOCOPhlAnSGB_202307").}
\item{For \verb{filtered_taxa<-}: }{A \code{list} with otu_table and tax_table (as matrices) of features that were filtered out.}
\item{For \verb{filters<-}: }{A named vector specifying used filter parameters (e.g., c(min_abundance = 0.01, min_prevalence = 0.5)).}
}}
}
\value{
\describe{
\item{For \code{info()}: }{A \code{list} containing metadata.}
\item{For \code{infoKeys()}: }{Lists valid keys for \code{microEDA} info slot.}
\item{For \code{infoFields()}: }{Lists available field(s) in \code{object} info slot.}
\item{For \code{taxrank()}: }{A \code{character} string with the lowest taxonomic rank.}
\item{For \code{transforms()}: }{A \code{character} vector of applied transformations.}
\item{For \code{mpa_version()}: }{A \code{character} string with the used the MetaPhlAn database version.}
\item{For \code{filtered_taxa()}: }{A \code{list} with otu_table and tax_table of features that were filtered out.}
\item{For \code{filters()}: }{A \code{list} of used filter parameters.}
}
Replacement methods (\verb{<-}) will return a \code{microEDA} object with the updated
corresponding slots.
}
\description{
Accessor methods for either retrieving or setting the info slot in a
\linkS4class{microEDA} object.
Additionally, each expected element of the info slot can be retrieved or set
individually.
Valid keys for the \code{microEDA} info slot can be displayed with \code{infoKeys()},
while filled info fields of a \code{microEDA} object can be displayed with
\code{infoFields()}.
}
\details{
\code{taxrank}: represents the lowest available taxonomic rank.

\code{transforms}: represents transformation(s) that were applied to the abundance data (otu_table).

\code{mpa_version}: represents the MetaPhlAn database version that was used to create the profile.

\code{filtered_taxa}: represents the otu_table and tax_table of features that were filtered out.

\code{filters}: represents filters that were applied to the abundance data (otu_table).
}
\seealso{
\code{\link{mpa_version,metaphlanProfile-method}}
}
