% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlanProfile-class.R
\name{load_metaphlan}
\alias{load_metaphlan}
\title{Load MetaPhlAn profile}
\usage{
load_metaphlan(file_path)
}
\arguments{
\item{file_path}{A \code{character} string containing the path to the file to read.}
}
\value{
An object of class \code{metaphlanProfile}.
}
\description{
Function to load a MetaPhlAn profile from a file path.
}
\examples{
file_path <- system.file("extdata", "merged_abundance_table.txt", package = "microEDA")

mtp <- load_metaphlan(file_path)
}
