% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microEDA-class.R
\docType{class}
\name{microEDA-class}
\alias{microEDA-class}
\alias{microEDA}
\alias{microEDA,metaphlanProfile-method}
\alias{microEDA,phyloseq-method}
\title{An S4 class extending phyloseq for microbiome data analysis}
\usage{
microEDA(tax_profile, metadata = NULL, sample_names_col = NULL)

\S4method{microEDA}{metaphlanProfile}(tax_profile, metadata = NULL, sample_names_col = NULL)

\S4method{microEDA}{phyloseq}(tax_profile, metadata = NULL, sample_names_col = NULL)
}
\arguments{
\item{tax_profile}{An object of class "metaphlanProfile" or "phyloseq".}

\item{metadata}{Optional data.frame with sample metadata.}

\item{sample_names_col}{Column name to use as sample identifiers.}
}
\value{
An object of class \code{microEDA}.
}
\description{
This class inherits from phyloseq and adds a slot for additional information,
such as lowest taxonomic rank, MetaPhlAn database version, applied
transformations and/or filters.
}
\section{Slots}{

\describe{
\item{\code{info}}{A list containing metadata.}
}}

\examples{
data("GlobalPatterns", package = "phyloseq")
me <- microEDA(GlobalPatterns)
}
