% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlanProfile-class.R
\docType{methods}
\name{mpa_taxProfile}
\alias{mpa_taxProfile}
\alias{mpa_taxProfile,metaphlanProfile-method}
\alias{mpa_taxProfile<-}
\alias{mpa_taxProfile<-,metaphlanProfile-method}
\title{Get or set the MetaPhlAn taxonomic profile}
\usage{
mpa_taxProfile(object)

mpa_taxProfile(object) <- value
}
\arguments{
\item{object}{An object of class \code{metaphlanProfile}.}

\item{value}{A \code{data.frame} with a MetaPhlAn taxonomic profile.}
}
\value{
The taxonomic profile (when getting), or updated \code{metaphlanProfile} (when setting).
}
\description{
Get or set the MetaPhlAn taxonomic profile
}
