% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlanProfile-class.R
\docType{methods}
\name{mpa_version}
\alias{mpa_version}
\alias{mpa_version,metaphlanProfile-method}
\alias{mpa_version<-}
\alias{mpa_version<-,metaphlanProfile-method}
\title{Get or set the MetaPhlAn database version.}
\usage{
mpa_version(object)

mpa_version(object) <- value
}
\arguments{
\item{object}{An object of class \code{metaphlanProfile}.}

\item{value}{A \code{character} string with a new MetaPhlAn database version (used only in assignment).}
}
\value{
The recorded MetaPhlAn database version (when getting), or updated \code{metaphlanProfile} (when setting).
}
\description{
Get or set the MetaPhlAn database version.
}
\seealso{
\code{\link{mpa_version,microEDA-method}}
}
